updated unit test
